ocr-service/requirements.txt
flask
pytesseract
Pillow
opencv-python
